Allow specifying multiple identities for a match
A list of identities can be specified in the second element of a match
tuple, and all the keys will be loaded into the agent for the first
identity in the list (the "primary" identity).
This is useful when you want to forward some keys only when connecting
to a specific host

If several identities have the same ssh-add options, all their keys will
be loaded with a single call to ssh-add.